requirements.txt
paho-mqtt
pygame
python-dotenv
pyinstaller
